Merged PR 6757: Deauthorize support for client SDK
Added DeauthorizeApplicationAsync to client connection, which deletes connection information and deauthorizes the app for all authorized records.
Changed GetConnection method to a synchronous one that does not try to run authentication, making it on-demand.
Updated PersonInfo property to GetPersonInfoAsync() to support idea that the connection object might not have the information on hand.
Moved SessionCredential and ServiceInfo properties off of the public IHealthVaultConnection to IConnectionInternal
Fixed bug in android browser auth broker that caused it to fail on the second attempt.
Fixed Delete method in UwpSecretStore.
Fixed build/publish configurations for SandboxUwp project.