Fixed multi-part requests on ThingClient (#39)
There were a couple problems: First there was a bug on ThingDeserializer where the connection passed in wasn't getting saved. Second, there were a couple of missing .ConfigureAwait(false) statements, meaning that the synchronous access in ThingCollection was causing a deadlock.
Also added a unit test to validate this scenario.